Renamed QRCodeScanner to QrCodeScanner
