+++ ui/src/lib/components/contacts/QrCodeScanner.svelte
